+++ driver/src/test/java/org/neo4j/driver/v1/integration/TemporalTypesIT.java
@@ -275,7 +275,7 @@ public class TemporalTypesIT
         assertEquals( value, converter.apply( record.get( 0 ) ) );
     }
 
-    private static IsoDuration newDuration( long months, long days, long seconds, long nanoseconds )
+    private static IsoDuration newDuration( long months, long days, long seconds, int nanoseconds )
     {
         return isoDuration( months, days, seconds, nanoseconds ).asIsoDuration();
     }
